requirements.txt
interface
requests
PyJWT
Django==2.0.2
django-webpack-loader==0.6.0
djangorestframework==3.7.7
mysql-connector-python==8.0.6
mysqlclient==1.3.12
pytz==2018.3
reverse==0.1.0
